Created a horizontal bar chart, a bubble chart & a guage chart
Used the D3 library to fetch data for plotting the charts.
Created a horizontal bar chart with a dropdown menu by slicing the top 10 OTUs found in that individual.
Created a bubble chart that displays each sample.
Created a guage chart that updates the chart whenever a new sample is selected.